AOSP VM Level Instrumentation: Compiled Code Support
Summary:
This diff adds support for compiled code instrumentation. Instrumentation overhead is 20-25 nanoseconds per method on a Nexus 6P for both 32 and 64-bit implementations.

**How it works**

An `HGraph` is the representation of the control flow of a method. The nodes in the graph are `HInstructions`. Here are some examples of existing `HInstructions`:

* `HNewInstance`: Represents a `new-instance` [[ https://source.android.com/devices/tech/dalvik/dalvik-bytecode | dex opcode ]].
* `HInvokeStaticOrDirect`: Represents either `invoke-direct` or `invoke-static`
* `HNativeDebugInfo`: Pseudo-instruction which provides the native debugger with mapping information.

In order to let the compiler know where to generate our tracing logic, we introduce two new `HInstructions`:

* `HTraceStart`: Pseudo-instruction that marks the start of the method trace.
* `HTraceEnd`: Pseudo-instruction that marks the end of the method trace.

In `instruction_builder.cc` we insert our new instructions into the graph. `HTraceStart` is added at the beginning of the method and `HTraceEnd` is added before every return. Now that the compiler knows about our new instructions, we just need to do the actual code generation in the corresponding `VisitTraceStart` and `VisitTraceEnd` methods. Here's some pseudocode that outlines the behavior we want:

```
VisitTraceStart:
    if traceEnabled(current_thread):
        data[i++] = method
        data[i++] = timestamp

VisitTraceEnd:
    if traceEnabled(current_thread):
        data[i++] = null // Use null to represent a stack pop
        data[i++] = timestamp
```
And here's the same thing in ARMv8-A pseudo-assembly for `VisitTraceStart`:
```
TR = thread register
MR = method register
data_ptr = temp register
timestamp = temp register

VisitTraceStart:
    LDR data_ptr, [TR, data_ptr_offset]  // data_ptr = thread.data_ptr
    CBZ data_ptr, done                   // if (data_ptr != null)
    MRS timestamp, SYS_CNTVCT_EL0        //     timestamp = <Generic Timer count>
    STR MR, [data_ptr], +8               //     *data_ptr++ = method
    STR timestamp, [data_ptr], +8        //     *data_ptr++ = timestamp
    STR data_ptr, [TR, data_ptr_offset]  //     *thread.data_ptr = data_ptr
  done:
```
We generate similar code in the ARMv7-A compiler to support apps that use the 32-bit instruction set (like our Rider app).

Test Plan: Ran 32-bit and 64-bit sample app and verified output file.

Reviewers: msw, attwell, adam.welc, marcello

Reviewed By: attwell

Subscribers: lukesc

Revert Plan: no

Differential Revision: https://code.uberinternal.com/D1438601