ART Tracing: App-writable System Property
Summary: The current system property we use to start and stop tracing is 'arttracing'. Unfortunately, this is only writable from the shell and not from an Android application (See AOSP/system/sepolicy/property_contexts). This change updates the property we use to 'dev.arttracing' which is writable by both the shell and system Android applications. We leave 'arttracing' as an alternate system property for backwards compatibility.

Test Plan: Tested that we're able to start and stop tracing by setting 'dev.arttracing' from an Android app.

Reviewers: attwell, msw

Reviewed By: attwell

Revert Plan: none

Differential Revision: https://code.uberinternal.com/D1529675