Remove hint that preoritizes blacklisting
**Background**
A branch misprediction is quite expensive. A mispredicted branch can easily be 10x more expensive than a correctly predicted branch.

So this compiler hint is likely doubling the cost of this method. Until we use blacklisting it doesn't make sense to include this. Unless we make big bets on blacklisting it might even be cheaper to remove support for blacklisting altoghether.