Analysis for impact of inlining and memory allocations on perf
We already know that the TraceStart() and TraceStop() calls add 20ms of overhead (on average) on a 6P since this was easy to measure directly. But there are other impacts that Nanoscope can have on performance: it allocates a big chunk of memory on startup and it disables inlining. This diff shows that the impact of these operations are minimal.

This diff modifies our Nanoscope build so that it re-enables inlining whenenver tracing is disables. Then, I used a script to alternate between running the app twice in a row without tracing and twice in a row with tracing. Between every run I clear the compilation cache.

**The conclusion**
Nanoscope doesn't have a significant impact beyond the 20ms overhead that we already know about. See results https://docs.google.com/spreadsheets/d/1yF_aOiQruMbOFXuJsKsfvO_BaXbc1zIzNnMKtxPklNQ/edit#gid=0 . The inlining itself doesn't appear to have any impact on performance. However, whether we allocate the 40MB buffer does appear to have some small impact on startup performance (around 10%). If we wanted to get more certain we could run this script overnight for 500 runs.